Convert to custom collector to avoid stale metrics
When the device was removed from the system smart_exporter still exposes
its metrics - the removed device appears in labels.

As far as I could research this is intended behavior for Prometheus
metrics - metrics label values are supposed to be static throughout the
lifetime of the binary. They can appear but the removing of the metric's
label values are not usual.

But in case of smart exporter this is not uncommon - when some SMART
alert fires up the disk will be replaced soon. Yet, the old metrics will
remain.

The correct way to make this work is to use so-called const metrics that
are generated on each scrape.

To expose const metrics we have to write custom collector by
implementing `Collector` interface.

This is what this commit does - replaces GaugeVec metrics with custom
type SMARTMetric that implements `Collector` interface. In `Collect`
method metrics values are created with `MustNewConstMetric`.

Fixes #1.